Fix warnings on 32 bit machines:
src/pusb.c: In function ‘send_file’:
src/pusb.c:342: warning: format ‘%ld’ expects type ‘long int’, but argument 2 has type ‘unsigned int’
src/pusb.c:354: warning: format ‘%ld’ expects type ‘long int’, but argument 2 has type ‘unsigned int’

on 32 bit Linux system.

+++ src/pusb.c
@@ -342,7 +342,7 @@ int send_file(struct usb_device *dev, char *f_name)
 		E_PRINT
 		    ("CMD:Expected to write %ld, "
 		     "actual write %d - err str: %s\n",
-		     sizeof(command), ret, strerror(errno));
+		     (long int)sizeof(command), ret, strerror(errno));
 		fail = -1;
 		goto closeup;
 	}
@@ -354,7 +354,7 @@ int send_file(struct usb_device *dev, char *f_name)
 		E_PRINT
 		    ("FSize:Expected to write %ld, "
 		     "actual write %d - err str: %s\n",
-		     sizeof(filesize), ret, strerror(errno));
+		     (long int)sizeof(filesize), ret, strerror(errno));
 		fail = -1;
 		goto closeup;
 	}
